+++ lib/json_api_client/middleware/http_client.ex
@@ -4,10 +4,18 @@ defmodule JsonApiClient.Middleware.HTTPClient do
   HTTP client Middleware based on HTTPoison library.
   """
 
-  alias JsonApiClient.{Response, RequestError}
+  alias JsonApiClient.{Response, RequestError, Request}
 
-  def call(%{method: method, url: url, body: body, headers: headers, http_options: http_options}, _, _) do
-    case HTTPoison.request(method, url, body, headers, http_options) do
+  def call(%Request{} = req, _, _) do
+    url          = Request.get_url(req)
+    headers      = req.headers
+                   |> Enum.into([])
+    http_options = req.options
+                   |> Map.put(:params, Request.get_query_params(req))
+                   |> Enum.into([])
+    body = Request.get_body(req)
+
+    case HTTPoison.request(req.method, url, body, headers, http_options) do
       {:ok, response} -> {:ok, %Response{
         status: response.status_code,
         headers: response.headers,