+++ CHANGELOG.md
@@ -4,9 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.2.0] 2017-10-27
+### Changed
+- Middleware call `request` parameter type is changed to `Request`. 
 
-## [1.1.0] - 2017-25-17
+### Added
+- Added `Request.attributes`
+- Added `DefaultRequestConfig` middleware
+- Added `Request.new/0` method 
+
+## [1.1.0] - 2017-10-25
 ### Changed
 - Renamed client_name to user_agent_suffix
 - use Mix.Project.config[:app] as a default value for user_agent_suffix